# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "description": "Mirror of `app.core.errors.error_envelope()` — the only error shape this API emits.\n\n`code` is typed as `ErrorCode` on purpose (design D11): it publishes the catalogue of\ncodes as an enum, so a consumer can switch over it exhaustively with its compiler's\nhelp. `app/core/error_codes.py` is the single source of truth that keeps that\ncatalogue from going stale.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "PropertyDetailResponse": {
        "description": "The aggregate of PRD §9.2 (`dto.ts:161-174`, R2.1).",
        "properties": {
          "access": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AccessResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "cleaning_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cleaning Status"
          },
          "current_or_next_reservation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReservationSummaryResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "financial": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FinancialSummaryResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "guest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GuestResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_cleaning_photos": {
            "items": {
              "$ref": "#/components/schemas/app__dashboard__api__schemas__CleaningPhotoResponse"
            },
            "title": "Last Cleaning Photos",
            "type": "array"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "open_incidents": {
            "items": {
              "$ref": "#/components/schemas/IncidentSummaryResponse"
            },
            "title": "Open Incidents",
            "type": "array"
          },
          "operational_state": {
            "$ref": "#/components/schemas/PropertyOperationalState"
          },
          "pending_approvals": {
            "items": {
              "$ref": "#/components/schemas/PendingApprovalResponse"
            },
            "title": "Pending Approvals",
            "type": "array"
          },
          "property_code": {
            "title": "Property Code",
            "type": "string"
          },
          "property_id": {
            "format": "uuid",
            "title": "Property Id",
            "type": "string"
          }
        },
        "required": [
          "property_id",
          "property_code",
          "operational_state",
          "current_or_next_reservation",
          "guest",
          "access",
          "cleaning_status",
          "last_cleaning_photos",
          "open_incidents",
          "financial",
          "notes",
          "pending_approvals"
        ],
        "title": "PropertyDetailResponse",
        "type": "object"
      },
      "PropertyOperationalState": {
        "enum": [
          "VACANT_READY",
          "AWAITING_CHECKIN",
          "OCCUPIED_ESTIMATED",
          "AWAITING_CLEANING",
          "CLEANING_SCHEDULED",
          "CLEANING_IN_PROGRESS",
          "READY_FOR_NEXT_GUEST",
          "MAINTENANCE_REQUIRED",
          "CRITICAL_INCIDENT",
          "BLOCKED_BY_OWNER",
          "OUT_OF_SERVICE"
        ],
        "title": "PropertyOperationalState",
        "type": "string"
      },
      "PropertyPageResponse": {
        "description": "The pagination envelope of PRD §23.",
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/PropertyResponse"
            },
            "title": "Data",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "per_page": {
            "title": "Per Page",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "per_page",
          "total_pages"
        ],
        "title": "PropertyPageResponse",
        "type": "object"
      },
...
      "PropertyStatus": {
        "enum": [
          "ACTIVE",
          "INACTIVE"
        ],
        "title": "PropertyStatus",
        "type": "string"
      },
...
      "UserRole": {
        "enum": [
          "SUPER_ADMIN",
          "TENANT_OWNER",
          "PROPERTY_MANAGER",
          "CLEANER",
          "TECHNICIAN"
        ],
        "title": "UserRole",
        "type": "string"
      },
      "UserStatus": {
        "enum": [
          "ACTIVE",
          "INACTIVE",
          "SUSPENDED"
        ],
        "title": "UserStatus",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "AutoHostAI backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/properties": {
      "get": {
        "description": "Paginated with `page`/`per_page` (PRD §23). Filters combine with AND. Ordered by name with the id as tiebreaker, so paging neither repeats a row nor skips one when two properties share a name. The wifi password is never in the response — `has_wifi_password` reports whether one is stored.",
        "operationId": "list_properties_api_v1_properties_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 100000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Per Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PropertyStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "current_operational_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/PropertyOperationalState"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Current Operational State"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing, malformed or expired credentials."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authenticated, but the role lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the tenant's properties",
        "tags": [
          "properties"
        ]
      },
      "post": {
        "description": "`internal_code` must be unique within the tenant and `pms_external_id` unique within the tenant AND the provider, both answering `409` on collision. `current_operational_state` cannot be chosen: a new property starts `VACANT_READY` and only `PropertyStateMachine` moves it. A `wifi_password` is encrypted before storage and can never be read back.",
        "operationId": "create_property_api_v1_properties_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePropertyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing, malformed or expired credentials."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authenticated, but the role lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Register a property",
        "tags": [
          "properties"
        ]
      }
    },
    "/api/v1/properties/{property_id}": {
      "get": {
        "description": "A property of another tenant answers `404`, with a body indistinguishable from one that does not exist. The wifi password is not returned in any form, masked included; `has_wifi_password` reports whether one is stored.",
        "operationId": "get_property_api_v1_properties__property_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Property Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing, malformed or expired credentials."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authenticated, but the role lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "One property",
        "tags": [
          "properties"
        ]
      },
      "patch": {
        "description": "Only the fields present in the body are applied. `current_operational_state` is not among them and is rejected with `422`. A body that changes nothing writes nothing and records no audit entry — except `wifi_password`, whose no-op cannot be detected because the stored value has no reader, so sending it always counts as a change. Retire a property with `{\"status\": \"INACTIVE\"}`; there is no `DELETE`.",
        "operationId": "update_property_api_v1_properties__property_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Property Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePropertyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing, malformed or expired credentials."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authenticated, but the role lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update a property partially",
        "tags": [
          "properties"
        ]
      }
    },
    "/api/v1/properties/{property_id}/dashboard": {
      "get": {
        "description": "The aggregate of PRD §9.2: reservation, guest, access, cleaning, incidents, financial, notes and pending approvals in one call. The guest is a name and the access a status label — never a document number, never an access code in any form, masked included. `last_cleaning_photos` is always empty until signed URLs exist; the blocks whose writing domain has not shipped yet (`incidents`, `owner_approvals`, `expenses`) query their real tables and come back empty, so the contract will not change when those changes land. `notes` is always `null` for now and deliberately so — no column owns it, and the candidates are free text an operator can paste a door code into. A property of another tenant answers `404`, indistinguishable from one that does not exist.",
        "operationId": "get_property_dashboard_api_v1_properties__property_id__dashboard_get",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Property Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing, malformed or expired credentials."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authenticated, but the role lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Everything happening on one property",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/api/v1/properties/{property_id}/state": {
      "get": {
        "description": "The light endpoint of PRD §23:1942, for refreshing an indicator without fetching the aggregate. Returns the canonical `PropertyOperationalState` literal — never translated — and the ISO-8601 UTC instant of the last transition, both **read** and neither recomputed: the state is whatever `PropertyStateMachine` last wrote. `last_transition_at` is `null` for a property that has never moved, because creation is not a transition. A property of another tenant answers `404`, indistinguishable from one that does not exist.",
        "operationId": "get_property_state_api_v1_properties__property_id__state_get",
        "parameters": [
          {
            "in": "path",
            "name": "property_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Property Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyStateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing, malformed or expired credentials."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authenticated, but the role lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "A property's operational state",
        "tags": [
          "properties"
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "description": "Paginated with `page`/`per_page` (PRD §23) and filterable by role and status, which combine with AND. Ordered by name with the id as a tiebreaker, so paginating never repeats or skips a row. Never includes any password material.",
        "operationId": "list_users_api_v1_users_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "maximum": 100000,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "per_page",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Per Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UserRole"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UserStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing, malformed or expired credentials."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authenticated, but the role lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List the tenant's users",
        "tags": [
          "users"
        ]
      },
      "post": {
        "description": "The password is generated by the system and returned **exactly once**, in this response; it is never stored in cleartext, logged, or written to the audit trail. Answers `409` if the address is already in use anywhere in the installation (ADR 0005), and `422` for `SUPER_ADMIN`, whose powers are global rather than tenant-scoped.",
        "operationId": "create_user_api_v1_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedUserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Missing, malformed or expired credentials."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Authenticated, but the role lacks the required permission."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create a user and issue its temporary password",
        "tags": [
          "users"
        ]
      }
    },
...
  }
}